fix router conflict
